alert 확인 취소 추가, 수정기능 권한 추가 작성자일 때만 수정 가능

+++ src/main/resources/static/js/board/board.js
@@ -3,21 +3,25 @@ const deleteButton = document.getElementById('delete-btn');
 
 if (deleteButton) {
     deleteButton.addEventListener('click', event => {
-        let id = document.getElementById('board-id').value;
+    // 사용자에게 확인을 받기 위한 경고창 표시
+    const isConfirmed = confirm('삭제 하시겠습니까?');
+        if (isConfirmed) {
+            let id = document.getElementById('board-id').value;
 
-        fetch(`/api/board/${id}`, {
-            method: 'DELETE'
-        })
-        .then(response => {
-            if (response.ok) {
-                // 서버에서 삭제가 성공적으로 이루어진 경우
-                alert('삭제 되었습니다.');
-                location.replace('/board');
-            } else {
-                // 서버에서 삭제가 실패한 경우 또는 권한이 없는 경우
-                alert('삭제 권한이 없습니다.');
-            }
-        })
+            fetch(`/api/board/${id}`, {
+                method: 'DELETE'
+            })
+            .then(response => {
+                if (response.ok) {
+                    // 서버에서 삭제가 성공적으로 이루어진 경우
+                    alert('삭제 되었습니다.');
+                    location.replace('/board');
+                } else {
+                    // 서버에서 삭제가 실패한 경우 또는 권한이 없는 경우
+                    alert('삭제 권한이 없습니다.');
+                }
+            })
+        }
     });
 }
 
@@ -29,26 +33,34 @@ const modifyButton = document.getElementById('modify-btn');
 if (modifyButton) {
     // 클릭 이벤트 발생시 수정 API 요청
     modifyButton.addEventListener('click', event => {
-        let params = new URLSearchParams(location.search);
-        let id = params.get('id');
+    const isConfirmed = confirm('수정 하시겠습니까?');
+        if (isConfirmed) {
+            let params = new URLSearchParams(location.search);
+            let id = params.get('id');
 
-        // 요청을 보낼 때 headers의 요청 형식을 지정
-        fetch(`/api/board/${id}`, {
-            method: 'PUT', // 수정 요청 메소드
-            headers: {
-                "Content-Type": "application/json",
-            },
+            // 요청을 보낼 때 headers의 요청 형식을 지정
+            fetch(`/api/board/${id}`, {
+                method: 'PUT', // 수정 요청 메소드
+                headers: {
+                    "Content-Type": "application/json",
+                },
 
-            // body에 HTML에 입력한 데이터를 JSON 형식으로 바꿔 보낸다.
-            body: JSON.stringify({
-                title: document.getElementById('title').value,
-                content: document.getElementById('content').value
+                // body에 HTML에 입력한 데이터를 JSON 형식으로 바꿔 보낸다.
+                body: JSON.stringify({
+                    title: document.getElementById('title').value,
+                    content: document.getElementById('content').value
+                })
             })
-        })
-        .then(() => {
-            alert('수정 되었습니다.');
-            location.replace(`/board/${id}`);
-        });
+            .then(response => {
+                if (response.ok) {
+                    alert('수정되었습니다.');
+                    location.replace(`/board/${id}`);
+                } else {
+                    // 403 Forbidden: 수정 권한이 없는 경우
+                    alert('수정 권한이 없습니다.');
+                }
+            })
+        }
     });
 }
 
@@ -59,23 +71,27 @@ const createButton = document.getElementById('create-btn');
 
 if (createButton) {
     createButton.addEventListener('click', event => {
-        // 등록 API로 POST 요청을 보낸다.
-        fetch(`/api/board`, {
-            method: 'POST',
-            headers: { // 요청을 보낼 때 headers의 요청 형식을 지정한다.
-                "Content-Type": "application/json",
-            },
 
-            // body에 HTML에서 입력한 데이터를 JSON 형식으로 바꿔 보낸다.
-            body: JSON.stringify({
-                title: document.getElementById('title').value,
-                content: document.getElementById('content').value,
-                author: document.getElementById('author').value
+    const isConfirmed = confirm('등록하시겠습니까?')
+        if (isConfirmed) {
+            // 등록 API로 POST 요청을 보낸다.
+            fetch(`/api/board`, {
+                method: 'POST',
+                headers: { // 요청을 보낼 때 headers의 요청 형식을 지정한다.
+                    "Content-Type": "application/json",
+                },
+
+                // body에 HTML에서 입력한 데이터를 JSON 형식으로 바꿔 보낸다.
+                body: JSON.stringify({
+                    title: document.getElementById('title').value,
+                    content: document.getElementById('content').value,
+                    author: document.getElementById('author').value
+                })
             })
-        })
-        .then(() => {
-            alert('등록 되었습니다.');
-            location.replace('/board');
-        });
+            .then(() => {
+                alert('등록 되었습니다.');
+                location.replace('/board');
+            });
+        }
     });
 }
